backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain-google-genai
langchain-core
langgraph
langchain-community
qdrant-client
python-dotenv
pdfplumber
